backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
numpy==1.26.3
torch==2.1.2
torchaudio==2.1.2
pyannote.audio==3.1.1
openai>=1.30.0
python-dotenv==1.0.0
scipy==1.12.0
httpx>=0.27.0,<0.28.0
openai-whisper==20231117
